refactor(timeline): extract presenter transport into an opt-in present-mode seam
Move PresenterTransportState, presenterAdvance/skip, applyPresenterCommandToMaster,
and wirePresenterCommands out of the always-on timeline composition path into
src/runtime/presenter-transport.ts. The transport is wired onto the master only
when a navigation forwards a presenter controller (mode=present); a non-present
navigation never instantiates it.

Keep the GSAP composition spine in timeline.ts (composeMasterTimeline, the scene
label namespace, assertSceneTimeline, MasterBeat). Factor a single validateLabelTime
helper, table-drive the pause-at-beat positionMaster branches, and fold the
onMaster/onSegmentChange observability into a SegmentReporter object.

Public signatures, data-pulsar-* attributes, and cross-engine timing behavior
unchanged.
